Extract optimistic response and mutation variables js_only
